+++ examples/svelte/vite.config.ts
@@ -2,8 +2,8 @@ import { svelte } from "@sveltejs/vite-plugin-svelte";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [svelte()],
-  optimizeDeps: {
-    exclude: ["@videts/vide"],
-  },
+	plugins: [svelte()],
+	optimizeDeps: {
+		exclude: ["@videts/vide"],
+	},
 });